fancy_regex
Struct Captures

pub struct Captures<'t> { /* private fields */ }

A set of capture groups found for a regex.

Implementations

impl<'t> Captures<'t>

pub fn get(&self, i: usize) -> Option<Match<'t>>

Get the capture group by its index in the regex.

If there is no match for that group or the index does not correspond to a group, None is returned. The index 0 returns the whole match.

pub fn name(&self, name: &str) -> Option<Match<'t>>

Returns the match for a named capture group. Returns None the capture group did not match or if there is no group with the given name.

pub fn expand(&self, replacement: &str, dst: &mut String)

Expands all instances of $group in replacement to the corresponding capture group name, and writes them to the dst buffer given.

group may be an integer corresponding to the index of the capture group (counted by order of opening parenthesis where \0 is the entire match) or it can be a name (consisting of letters, digits or underscores) corresponding to a named capture group.

If group isn’t a valid capture group (whether the name doesn’t exist or isn’t a valid index), then it is replaced with the empty string.

The longest possible name is used. e.g., $1a looks up the capture group named 1a and not the capture group at index 1. To exert more precise control over the name, use braces, e.g., ${1}a.

To write a literal $, use $$.

For more control over expansion, see Expander.

pub fn iter<'c>(&'c self) -> SubCaptureMatches<'c, 't>

Iterate over the captured groups in order in which they appeared in the regex. The first capture corresponds to the whole match.

pub fn len(&self) -> usize

How many groups were captured. This is always at least 1 because group 0 returns the whole match.